// src/v6y-commons/src/types/DependencyStatusHelpType.d.ts
import { LinkType } from './LinkType.ts';
export interface DependencyStatusHelpType {
    _id?: number;
    category?: string;
    title?: string;
    description?: string;
    links?: LinkType[] | undefined;
}
export interface DependencyStatusHelpInputType {
    _id?: number;
    category?: string | undefined;
    title?: string;
    description?: string;
    links?: string[] | undefined;
}